pyproject.toml
[tool.poetry]
name = "django-indexer"
version = "0.1.0"
description = "Potlock indexer built using Django & NEAR Data Lake Framework"
authors = ["Lachlan Glen <54282009+lachlanglen@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
near-lake-framework = "^0.0.7"
celery = "^5.3.6"
redis = { extras = ["hiredis"], version = "^5.0.3" }
requests = "^2.31.0"
djangorestframework = "^3.15.1"
django-cachalot = "^2.6.2"
django-redis = "^5.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
